Float wraps double instead of big decimal; precision fixed; mandelbrot example uses floating point now

+++ tools/src/wyvern/stdlib/support/Float.java
@@ -3,7 +3,7 @@ package wyvern.stdlib.support;
 public class Float {
     public static final Float utils = new Float();
 
-    public float from(String s) {
-        return java.lang.Float.parseFloat(s.trim());
+    public double from(String s) {
+        return java.lang.Double.parseDouble(s.trim());
     }
 }
